requirements.txt
greenlet==3.3.1
packaging==26.0
playwright==1.58.0
pyee==13.0.0
python-dotenv==1.2.1
typing_extensions==4.15.0
